% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean.R
\name{mymean}
\alias{mymean}
\title{Alternative mean}
\usage{
mymean(x, trim = 0, graph = "NULL")
}
\arguments{
\item{x}{a numeric vector}

\item{trim}{value between 0 and 0.5}

\item{graph}{box, hist, or both}
}
\value{
A graph, list with arithmetic mean, and summary and a trimmed data csv file
}
\description{
Alternative mean
}
\examples{
\dontrun{mymean(x = rnorm(30), trim = 0.1, graph = "both")}
}
